requirements.txt
yt-dlp
pytube
pycryptodome
requests
m3u8
selenium
gputil